Don't let D::S::Abstract read/write cookies.
If using Dancer::Session::PSGI, PSGI should be handling all
the cookie management for the session, but the design of
Dancer::Session assumes it does the cookie management.
Overload some of the methods in Dancer::Session::Abstract
that we wouldn't normally overload to prevent Dancer::Session
from managing the cookies directly.